Address further comments
